# bidservice/src/main/java/com/crowd/auction/bidservice/dto/BidRequestDTO.java
package com.crowd.auction.bidservice.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BidRequestDTO {
    private String userId;
    private BigDecimal amount;
}