# bidservice/src/main/java/com/crowd/auction/bidservice/dto/BidResponseDTO.java
package com.crowd.auction.bidservice.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BidResponseDTO {
    private Long id;
    private Long itemId;
    private String userId;
    private BigDecimal amount;
    private LocalDateTime createdAt;
}