# bidservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Bid Service API
  description: API for placing and viewing bids on items in open auctions
  version: 0.0.1
servers:
  - url: http://localhost:8082
    description: Local Bid Service
  - url: http://localhost:8080/bidservice
    description: API Gateway
paths:
  /api/item/{id}/bid:
    post:
      summary: Place a bid on an item
      description: Placed bid must be higher than the current highest bid, and the auction must be active.
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidRequest'
      responses:
        '200':
          description: Bid placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponse'
        '400':
          description: Invalid bid (e.g. amount too low or auction closed)
        '404':
          description: Item not found
    get:
      summary: Get the last 20 bids for an item
      description: Returns the 20 most recent bids placed on the item, sorted by creation time descending.
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BidResponse'
components:
  schemas:
    BidRequest:
      type: object
      required:
        - userId
        - amount
      properties:
        userId:
          type: string
          example: user123
        amount:
          type: number
          format: double
          example: 150.00
    BidResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        itemId:
          type: integer
          format: int64
          example: 5
        userId:
          type: string
          example: user123
        amount:
          type: number
          format: double
          example: 150.00
        createdAt:
          type: string
          format: date-time
          example: '2026-06-12T18:00:00'